+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * :hammer: Premise's requireNonNull method rename to requireNotNull
 * :hammer: Premise's checkInRange and checkNotInRange method rename to requireInRange and requireNotInRange
 * :sparkles: Premise add check and checkNotNull method
+* :sparkles: Premise add requireSafe and requireNotSafe method
 
 ## 0.6.4
 * Char add isNotBlank, isBlankOr, isNotChinese, isNotChineseOr, isDigit, isNotDigit, isNotDigitOr,